+++ src/civilization/game_moves/game_moves_events.rs
@@ -1,0 +1,14 @@
+use bevy::prelude::{Entity, Event};
+
+#[derive(Event, Debug)]
+pub struct RecalculatePlayerMoves {
+    pub player: Entity,
+}
+
+impl RecalculatePlayerMoves {
+    pub fn new(player: Entity) -> Self {
+        RecalculatePlayerMoves {
+            player
+        }
+    }
+}
