+++ src/civilization/game_moves/mod.rs
@@ -1,1 +1,5 @@
 pub mod game_moves_plugin;
+pub mod game_moves_triggers;
+pub mod game_moves_events;
+pub mod game_moves_components;
+pub mod game_moves_systems;
